requirements.txt
youtube-transcript-api
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
requests
faster-whisper
yt-dlp